refactor SelectCalendarModal
