+++ ts/screens/profile/CalendarsPreferencesScreen.tsx
@@ -1,52 +1,21 @@
-import { Button, List, Text } from "native-base";
 import * as React from "react";
-import RNCalendarEvents, { Calendar } from "react-native-calendar-events";
+import { Calendar } from "react-native-calendar-events";
 import { NavigationInjectedProps } from "react-navigation";
 import { connect } from "react-redux";
-
-import * as pot from "italia-ts-commons/lib/pot";
+import CalendarsListContainer from "../../components/CalendarsListContainer";
 import LoadingSpinnerOverlay from "../../components/LoadingSpinnerOverlay";
-import { EdgeBorderComponent } from "../../components/screens/EdgeBorderComponent";
-import ListItemComponent from "../../components/screens/ListItemComponent";
-
 import ScreenContent from "../../components/screens/ScreenContent";
 import TopScreenComponent from "../../components/screens/TopScreenComponent";
 import I18n from "../../i18n";
 import { preferredCalendarSaveSuccess } from "../../store/actions/persistedPreferences";
-import { Dispatch, ReduxProps } from "../../store/actions/types";
-import { GlobalState } from "../../store/reducers/types";
+import { Dispatch } from "../../store/actions/types";
 
 type OwnProps = NavigationInjectedProps;
 
-type Props = ReturnType<typeof mapStateToProps> &
-  ReturnType<typeof mapDispatchToProps> &
-  ReduxProps &
-  OwnProps;
+type Props = ReturnType<typeof mapDispatchToProps> & OwnProps;
 
 type State = {
-  calendars: pot.Pot<ReadonlyArray<Calendar>, ResourceError>;
-};
-
-const INITIAL_STATE: State = {
-  calendars: pot.none
-};
-
-type FetchError = {
-  kind: "FETCH_ERROR";
-};
-
-type ResourceError = FetchError;
-
-const mapResourceErrorToMessage = (resourceError: ResourceError): string => {
-  switch (resourceError.kind) {
-    case "FETCH_ERROR":
-      return I18n.t("messages.cta.errors.fetchCalendars");
-
-    default: {
-      // Exhaustive check
-      return ((): never => resourceError.kind)();
-    }
-  }
+  isLoading: boolean;
 };
 
 /**
@@ -55,7 +24,9 @@ const mapResourceErrorToMessage = (resourceError: ResourceError): string => {
 class CalendarsPreferencesScreen extends React.PureComponent<Props, State> {
   constructor(props: Props) {
     super(props);
-    this.state = INITIAL_STATE;
+    this.state = {
+      isLoading: true
+    };
   }
 
   private onCalendarSelected = (calendar: Calendar) => {
@@ -63,83 +34,32 @@ class CalendarsPreferencesScreen extends React.PureComponent<Props, State> {
     this.props.navigation.goBack();
   };
 
-  public render() {
-    const { calendars } = this.state;
-    const { defaultCalendar } = this.props;
+  private onCalendarLoaded = () => {
+    this.setState({ isLoading: false });
+  };
 
+  public render() {
+    const { isLoading } = this.state;
     return (
-      <LoadingSpinnerOverlay isLoading={pot.isLoading(calendars)}>
+      <LoadingSpinnerOverlay isLoading={isLoading}>
         <TopScreenComponent
           headerTitle={I18n.t("profile.preferences.title")}
-          title={I18n.t("profile.preferences.list.preferred_calendar.title")}
           goBack={this.props.navigation.goBack}
         >
           <ScreenContent
             title={I18n.t("profile.preferences.list.preferred_calendar.title")}
             subtitle={I18n.t("messages.cta.reminderCalendarSelect")}
           >
-            {pot.isError(calendars) && (
-              <React.Fragment>
-                <Text>{mapResourceErrorToMessage(calendars.error)}</Text>
-                <Button onPress={this.fetchCalendars}>
-                  <Text>{I18n.t("global.buttons.retry")}</Text>
-                </Button>
-              </React.Fragment>
-            )}
-            {pot.isSome(calendars) && (
-              <List withContentLateralPadding={true}>
-                {calendars.value.map(calendar => {
-                  const isDefaultCalendar =
-                    defaultCalendar && calendar.id === defaultCalendar.id;
-                  return (
-                    <ListItemComponent
-                      key={calendar.id}
-                      title={calendar.title}
-                      hideIcon={!isDefaultCalendar}
-                      iconSize={12}
-                      iconName={isDefaultCalendar ? "io-tick-big" : undefined}
-                      onPress={() => this.onCalendarSelected(calendar)}
-                    />
-                  );
-                })}
-                <EdgeBorderComponent />
-              </List>
-            )}
+            <CalendarsListContainer
+              onCalendarSelected={this.onCalendarSelected}
+              onCalendarLoaded={this.onCalendarLoaded}
+            />
           </ScreenContent>
         </TopScreenComponent>
       </LoadingSpinnerOverlay>
     );
   }
-
-  public componentDidMount() {
-    this.fetchCalendars();
-  }
-
-  private fetchCalendars = () => {
-    this.setState({ calendars: pot.noneLoading });
-    // Fetch user calendars.
-    RNCalendarEvents.findCalendars()
-      .then(calendars => {
-        // Filter out only calendars that allow modifications
-        const editableCalendars = calendars.filter(
-          calendar => calendar.allowsModifications
-        );
-        this.setState({ calendars: pot.some(editableCalendars) });
-      })
-      .catch(_ => {
-        const fetchError: FetchError = {
-          kind: "FETCH_ERROR"
-        };
-        this.setState({
-          calendars: pot.toError(pot.none, fetchError)
-        });
-      });
-  };
 }
-
-const mapStateToProps = (state: GlobalState) => ({
-  defaultCalendar: state.persistedPreferences.preferredCalendar
-});
 
 const mapDispatchToProps = (dispatch: Dispatch) => ({
   preferredCalendarSaveSuccess: (calendar: Calendar) =>
@@ -151,6 +71,6 @@ const mapDispatchToProps = (dispatch: Dispatch) => ({
 });
 
 export default connect(
-  mapStateToProps,
+  undefined,
   mapDispatchToProps
 )(CalendarsPreferencesScreen);
